+++ convex/_generated/api.d.ts
@@ -50,9 +50,9 @@ import type * as util_asyncMap from "../util/asyncMap.js";
 import type * as util_compression from "../util/compression.js";
 import type * as util_geometry from "../util/geometry.js";
 import type * as util_isSimpleObject from "../util/isSimpleObject.js";
+import type * as util_llm from "../util/llm.js";
 import type * as util_minheap from "../util/minheap.js";
 import type * as util_object from "../util/object.js";
-import type * as util_openai from "../util/openai.js";
 import type * as util_sleep from "../util/sleep.js";
 import type * as util_types from "../util/types.js";
 import type * as util_xxhash from "../util/xxhash.js";
@@ -103,9 +103,9 @@ declare const fullApi: ApiFromModules<{
   "util/compression": typeof util_compression;
   "util/geometry": typeof util_geometry;
   "util/isSimpleObject": typeof util_isSimpleObject;
+  "util/llm": typeof util_llm;
   "util/minheap": typeof util_minheap;
   "util/object": typeof util_object;
-  "util/openai": typeof util_openai;
   "util/sleep": typeof util_sleep;
   "util/types": typeof util_types;
   "util/xxhash": typeof util_xxhash;
