+++ convex/init.ts
@@ -7,7 +7,7 @@ import { insertInput } from './aiTown/insertInput';
 import { Id } from './_generated/dataModel';
 import { createEngine } from './aiTown/main';
 import { ENGINE_ACTION_DURATION } from './constants';
-import { assertOpenAIKey } from './util/openai';
+import { assertOpenAIKey } from './util/llm';
 
 const init = mutation({
   args: {
